# backend/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "title": "CRM Backend API",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
        "/auth/login": {
            "post": {
                "description": "Логин",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "аутентификация"
                ],
                "summary": "Логин",
                "parameters": [
                    {
                        "description": "Тело запроса",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Login"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/response.Login"
                        }
                    }
                }
            }
        },
        "/auth/refresh": {
            "post": {
                "description": "Регенерация токена",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "аутентификация"
                ],
                "summary": "Регенерация токена",
                "parameters": [
                    {
                        "description": "Тело запроса",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Refresh"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Refresh"
                        }
                    }
                }
            }
        },
        "/v1/hotel/rooms": {
            "post": {
                "description": "Получение доступных типов комнат",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "отель"
                ],
                "summary": "Получение доступных типов комнат",
                "parameters": [
                    {
                        "description": "Тело запроса",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.GetRooms"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.GetAllRooms"
                        }
                    }
                }
            }
        },
        "/v1/hotel/rooms/dump_token": {
            "get": {
                "description": "Получение токена комнаты [ХАРДКОД]",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "отель"
                ],
                "summary": "Получение токена комнаты [ХАРДКОД]",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Token"
                        }
                    }
                }
            }
        },
        "/v1/hotel/rooms/light": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Тогл света в комнате",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "отель"
                ],
                "summary": "Свет",
                "parameters": [
                    {
                        "description": "Тело запроса",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.Light"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/v1/hotel/rooms/reservations/confirmed": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получение бронирований, до заезда клиента",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "отель"
                ],
                "summary": "Получение бронирований, до заезда клиента",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/entity.ReservationInfo"
                            }
                        }
                    }
                }
            }
        },
        "/v1/hotel/rooms/reservations/{id}": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Обновление статуса брони",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "отель"
                ],
                "summary": "Обновление статуса брони",
                "parameters": [
                    {
                        "type": "string",
                        "description": "ID брони",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Тело запроса",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.UpdateReservationStatus"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/v1/hotel/rooms/reserve": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Бронь комнаты",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "отель"
                ],
                "summary": "Бронь",
                "parameters": [
                    {
                        "description": "Тело запроса",
                        "name": "input",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/request.ReserveRoom"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No Content"
                    }
                }
            }
        },
        "/v1/hotel/rooms/token": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получение токена комнаты",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "отель"
                ],
                "summary": "Получение токена комнаты",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.Token"
                        }
                    }
                }
            }
        },
        "/v1/hotel/rooms/{id}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Получение информации по комнате для админки",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "отель"
                ],
                "summary": "Получение информации по комнате для админки",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Идентификатор комнаты",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/response.RoomInfo"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "entity.ReservationInfo": {
            "type": "object",
            "properties": {
                "check_in": {
                    "type": "string"
                },
                "check_out": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "login": {
                    "type": "string"
                },
                "room_id": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "user_id": {
                    "type": "integer"
                }
            }
        },
        "entity.Room": {
            "type": "object",
            "properties": {
                "floor": {
                    "type": "integer"
                },
                "hotel_id": {
                    "type": "integer"
                },
                "room_number": {
                    "type": "string"
                },
                "room_type_id": {
                    "type": "integer"
                }
            }
        },
        "entity.RoomType": {
            "type": "object",
            "properties": {
                "capacity": {
                    "type": "integer"
                },
                "id": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                }
            }
        },
        "request.GetRooms": {
            "type": "object",
            "required": [
                "end_date",
                "start_date"
            ],
            "properties": {
                "end_date": {
                    "type": "string",
                    "example": "29.05.2026"
                },
                "start_date": {
                    "type": "string",
                    "example": "01.12.2025"
                }
            }
        },
        "request.Light": {
            "type": "object",
            "required": [
                "state"
            ],
            "properties": {
                "state": {
                    "type": "string",
                    "example": "true"
                }
            }
        },
        "request.Login": {
            "type": "object",
            "required": [
                "login"
            ],
            "properties": {
                "login": {
                    "type": "string",
                    "example": "admin"
                },
                "password": {
                    "type": "string",
                    "maxLength": 50,
                    "minLength": 4,
                    "example": "admin"
                }
            }
        },
        "request.Refresh": {
            "type": "object",
            "required": [
                "refresh_token"
            ],
            "properties": {
                "refresh_token": {
                    "type": "string",
                    "example": "refresh_token"
                }
            }
        },
        "request.ReserveRoom": {
            "type": "object",
            "required": [
                "end_date",
                "room_type",
                "start_date"
            ],
            "properties": {
                "end_date": {
                    "type": "string",
                    "example": "29.05.2026"
                },
                "room_type": {
                    "type": "integer",
                    "example": 1
                },
                "start_date": {
                    "type": "string",
                    "example": "01.12.2025"
                }
            }
        },
        "request.UpdateReservationStatus": {
            "type": "object",
            "required": [
                "status"
            ],
            "properties": {
                "status": {
                    "description": "confirmed, checked_in, checked_out",
                    "type": "string",
                    "example": "confirmed"
                }
            }
        },
        "response.GetAllRooms": {
            "type": "object",
            "properties": {
                "rooms": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/entity.RoomType"
                    }
                }
            }
        },
        "response.Login": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "access_token"
                },
                "refresh_token": {
                    "type": "string",
                    "example": "refresh_token"
                },
                "user": {
                    "$ref": "#/definitions/response.User"
                }
            }
        },
        "response.Refresh": {
            "type": "object",
            "properties": {
                "access_token": {
                    "type": "string",
                    "example": "access_token"
                },
                "refresh_token": {
                    "type": "string",
                    "example": "refresh_token"
                }
            }
        },
        "response.RoomInfo": {
            "type": "object",
            "properties": {
                "room": {
                    "$ref": "#/definitions/entity.Room"
                },
                "status": {
                    "type": "string"
                }
            }
        },
        "response.Token": {
            "type": "object",
            "properties": {
                "token": {
                    "type": "string"
                }
            }
        },
        "response.User": {
            "type": "object",
            "properties": {
                "business_id": {
                    "type": "integer"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "integer"
                },
                "login": {
                    "type": "string"
                },
                "user_type": {
                    "type": "string"
                }
            }
        }
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}